requirements.txt
asgiref==3.8.1
asttokens==2.4.1
black==24.10.0
blinker==1.8.2
click==8.1.7
decorator==5.1.1
Django==5.1.2
djangorestframework==3.15.2
executing==2.1.0
ipython==8.29.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
numpy==2.1.2
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
patsy==0.5.6
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
scipy==1.14.1
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
statsmodels==0.14.4
traitlets==5.14.3
tzdata==2024.2
wcwidth==0.2.13
Werkzeug==3.0.6
